ssh-plus: allow aliased names in compile/auth
documentation coming up in a future commit :) but basically, it allows
lines like this:

    useralias   foo = bar baz

where gitolite's access control (the R/RW/RW+ stuff) only knows about
"foo", but the external entity invoking gitolite's auth command
(currently ssh, but the plan is there will be others) may be using "bar"
or "baz" to refer to the **same** user (foo).

The bulk of the code is sanitychecks on this (making sure that foo does
not appear on the right side in some other useralias statement, to take
one example).