ssh-plus: make error reporting a little more flexible
Till now we used to treat most errors as pure warnings (life goes on) or
"die" (immediate exit).

The useralias logic is sort of in between -- there are likely to be
multiple errors in a large-ish config, so you want to show the user as
many of them as possible in one "compile", yet treating them as warnings
is too liberal.

Fixed that up with an error counter.