in addition to NAME/, also add SIZE/ and COUNT/ restrictions
They work somewhat differently than normal refexs.  The syntax is the
same but they "match" when the actual file being updated is smaller than
the limit in SIZE/ (or the number of files being updated is lesser than
the limit in COUNT/).

see conf/example.conf for a quick example.

----

PS: ladies and gentlemen, please welcome gitoheavy!  damn... I need to
stop listening to people who ask for more and more features :/